+++ packages/manager/apps/pci-ssh-keys/src/routes.tsx
@@ -1,18 +1,16 @@
 import { getProjectQuery } from '@ovh-ux/manager-pci-common';
 import queryClient from '@/queryClient';
 
-const lazyRouteConfig = (importFn: CallableFunction) => {
-  return {
-    lazy: async () => {
-      const { default: moduleDefault, ...moduleExports } = await importFn();
+const lazyRouteConfig = (importFn: CallableFunction) => ({
+  lazy: async () => {
+    const { default: moduleDefault, ...moduleExports } = await importFn();
 
-      return {
-        Component: moduleDefault,
-        ...moduleExports,
-      };
-    },
-  };
-};
+    return {
+      Component: moduleDefault,
+      ...moduleExports,
+    };
+  },
+});
 
 export default [
   {
@@ -22,9 +20,8 @@ export default [
   {
     id: 'ssh',
     path: '/pci/projects/:projectId/ssh',
-    loader: async ({ params }) => {
-      return queryClient.fetchQuery(getProjectQuery(params.projectId));
-    },
+    loader: async ({ params }) =>
+      queryClient.fetchQuery(getProjectQuery(params.projectId)),
     ...lazyRouteConfig(() => import('@/pages/Layout')),
     children: [
       {