+++ packages/manager/modules/pci/src/projects/project/storages/containers/container/data-encryption/component.js
@@ -6,6 +6,7 @@ const component = {
     encryption: '=',
     encryptionAlgorithms: '<',
     trackEncryptionAction: '<',
+    containerRegion: '<?',
   },
   template,
   controller,