+++ packages/manager/modules/pci/src/projects/project/storages/containers/containers.controller.js
@@ -1,7 +1,12 @@
 import find from 'lodash/find';
 import { getCriteria } from '../../project.utils';
 
-import { OBJECT_CONTAINER_OFFERS_TYPES } from './containers.constants';
+import {
+  OBJECT_CONTAINER_OFFERS_TYPES,
+  OBJECT_CONTAINER_MODE_MONO_ZONE,
+  OBJECT_CONTAINER_MODE_MULTI_ZONES,
+  OBJECT_CONTAINER_DEPLOYMENT_MODE_TYPES,
+} from './containers.constants';
 
 export default class PciStoragesContainersController {
   /* @ngInject */
@@ -17,6 +22,8 @@ export default class PciStoragesContainersController {
     this.publicToggleLoading = false;
     this.OBJECT_CONTAINER_OFFERS_TYPES = OBJECT_CONTAINER_OFFERS_TYPES;
     this.PciProjectStorageContainersService = PciProjectStorageContainersService;
+    this.OBJECT_CONTAINER_MODE_MONO_ZONE = OBJECT_CONTAINER_MODE_MONO_ZONE;
+    this.OBJECT_CONTAINER_MODE_MULTI_ZONES = OBJECT_CONTAINER_MODE_MULTI_ZONES;
   }
 
   $onInit() {
@@ -41,6 +48,9 @@ export default class PciStoragesContainersController {
       this.trackPage(this.trackingTag);
     }
     this.setContainerLoadingErrors();
+
+    this.setDeploymentModeOptions();
+    this.addDeploymentMode();
   }
 
   setSolutionTypeOptions() {
@@ -52,6 +62,27 @@ export default class PciStoragesContainersController {
         `pci_projects_project_storages_containers_offer_${type}`,
       );
     });
+  }
+
+  setDeploymentModeOptions() {
+    this.deploymentModeOptions = {
+      values: {},
+    };
+    Object.values(OBJECT_CONTAINER_DEPLOYMENT_MODE_TYPES).forEach((type) => {
+      this.deploymentModeOptions.values[type] = this.$translate.instant(
+        `pci_projects_project_storages_containers_deployment_mode_${type}`,
+      );
+    });
+  }
+
+  addDeploymentMode() {
+    this.containers = this.containers.reduce((acc, container) => {
+      acc.push({
+        ...container,
+        deploymentMode: this.regions[container.region].type,
+      });
+      return acc;
+    }, []);
   }
 
   addContainerSolution() {