+++ packages/manager/apps/pci-ssh-keys/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.10](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-ssh-keys-app@0.6.9...@ovh-ux/manager-pci-ssh-keys-app@0.6.10) (2024-10-08)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-ssh-keys-app
+
+
+
+
+
+## [0.6.9](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-ssh-keys-app@0.6.8...@ovh-ux/manager-pci-ssh-keys-app@0.6.9) (2024-10-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-ssh-keys-app
+
+
+
+
+
 ## [0.6.8](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-ssh-keys-app@0.6.7...@ovh-ux/manager-pci-ssh-keys-app@0.6.8) (2024-10-04)
 
 **Note:** Version bump only for package @ovh-ux/manager-pci-ssh-keys-app