+++ packages/manager/apps/pci/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.6.26](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.6.25...@ovh-ux/manager-pci-app@4.6.26) (2024-10-08)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
+## [4.6.25](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.6.24...@ovh-ux/manager-pci-app@4.6.25) (2024-10-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
 ## [4.6.24](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.6.23...@ovh-ux/manager-pci-app@4.6.24) (2024-10-03)
 
 **Note:** Version bump only for package @ovh-ux/manager-pci-app