+++ packages/manager/modules/iam/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.12.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-iam@1.11.3...@ovh-ux/manager-iam@1.12.0) (2024-10-07)
+
+
+### Features
+
+* **iam:** need disabled indicators ([#12698](https://github.com/ovh/manager/issues/12698)) ([cb4ee07](https://github.com/ovh/manager/commit/cb4ee0735695d776fd9d9c6a10738da498198d13))
+
+
+
+
+
 ## [1.11.3](https://github.com/ovh/manager/compare/@ovh-ux/manager-iam@1.11.2...@ovh-ux/manager-iam@1.11.3) (2024-08-29)