+++ packages/manager/modules/pci-universe-components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.23](https://github.com/ovh/manager/compare/@ovh-ux/ng-ovh-pci-universe-components@1.1.22...@ovh-ux/ng-ovh-pci-universe-components@1.1.23) (2024-10-08)
+
+**Note:** Version bump only for package @ovh-ux/ng-ovh-pci-universe-components
+
+
+
+
+
 ## [1.1.22](https://github.com/ovh/manager/compare/@ovh-ux/ng-ovh-pci-universe-components@1.1.21...@ovh-ux/ng-ovh-pci-universe-components@1.1.22) (2024-10-03)
 
 **Note:** Version bump only for package @ovh-ux/ng-ovh-pci-universe-components